Command reorg pt2
