+++ src/game/ChatCommands/Level3.cpp
@@ -973,35 +973,6 @@ bool ChatHandler::HandleReloadDisablesCommand(char * /*args*/)
     return true;
 }
 
-bool ChatHandler::HandleLoadScriptsCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    switch (sScriptMgr.LoadScriptLibrary(args))
-    {
-        case SCRIPT_LOAD_OK:
-            sWorld.SendWorldText(LANG_SCRIPTS_RELOADED_ANNOUNCE);
-            SendSysMessage(LANG_SCRIPTS_RELOADED_OK);
-            break;
-        case SCRIPT_LOAD_ERR_NOT_FOUND:
-            SendSysMessage(LANG_SCRIPTS_NOT_FOUND);
-            break;
-        case SCRIPT_LOAD_ERR_WRONG_API:
-            SendSysMessage(LANG_SCRIPTS_WRONG_API);
-            break;
-        case SCRIPT_LOAD_ERR_OUTDATED:
-            SendSysMessage(LANG_SCRIPTS_OUTDATED);
-            break;
-    }
-
-    return true;
-}
-
-
-
 bool ChatHandler::HandleMaxSkillCommand(char* /*args*/)
 {
     Player* SelectedPlayer = getSelectedPlayer();
@@ -3433,164 +3404,6 @@ bool ChatHandler::HandleQuestCompleteCommand(char* args)
     player->CompleteQuest(entry, QUEST_STATUS_FORCE_COMPLETE);
     return true;
 }
-
-bool ChatHandler::HandlePDumpLoadCommand(char* args)
-{
-    char* file = ExtractQuotedOrLiteralArg(&args);
-    if (!file)
-    {
-        return false;
-    }
-
-    std::string account_name;
-    uint32 account_id = ExtractAccountId(&args, &account_name);
-    if (!account_id)
-    {
-        return false;
-    }
-
-    char* name_str = ExtractLiteralArg(&args);
-
-    uint32 lowguid = 0;
-    std::string name;
-
-    if (name_str)
-    {
-        name = name_str;
-        // normalize the name if specified and check if it exists
-        if (!normalizePlayerName(name))
-        {
-            PSendSysMessage(LANG_INVALID_CHARACTER_NAME);
-            SetSentErrorMessage(true);
-            return false;
-        }
-
-        if (ObjectMgr::CheckPlayerName(name, true) != CHAR_NAME_SUCCESS)
-        {
-            PSendSysMessage(LANG_INVALID_CHARACTER_NAME);
-            SetSentErrorMessage(true);
-            return false;
-        }
-
-        if (*args)
-        {
-            if (!ExtractUInt32(&args, lowguid))
-            {
-                return false;
-            }
-
-            if (!lowguid)
-            {
-                PSendSysMessage(LANG_INVALID_CHARACTER_GUID);
-                SetSentErrorMessage(true);
-                return false;
-            }
-
-            ObjectGuid guid = ObjectGuid(HIGHGUID_PLAYER, lowguid);
-
-            if (sObjectMgr.GetPlayerAccountIdByGUID(guid))
-            {
-                PSendSysMessage(LANG_CHARACTER_GUID_IN_USE, lowguid);
-                SetSentErrorMessage(true);
-                return false;
-            }
-        }
-    }
-
-    switch (PlayerDumpReader().LoadDump(file, account_id, name, lowguid))
-    {
-        case DUMP_SUCCESS:
-            PSendSysMessage(LANG_COMMAND_IMPORT_SUCCESS);
-            break;
-        case DUMP_FILE_OPEN_ERROR:
-            PSendSysMessage(LANG_FILE_OPEN_FAIL, file);
-            SetSentErrorMessage(true);
-            return false;
-        case DUMP_FILE_BROKEN:
-            PSendSysMessage(LANG_DUMP_BROKEN, file);
-            SetSentErrorMessage(true);
-            return false;
-        case DUMP_TOO_MANY_CHARS:
-            PSendSysMessage(LANG_ACCOUNT_CHARACTER_LIST_FULL, account_name.c_str(), account_id);
-            SetSentErrorMessage(true);
-            return false;
-        default:
-            PSendSysMessage(LANG_COMMAND_IMPORT_FAILED);
-            SetSentErrorMessage(true);
-            return false;
-    }
-
-    return true;
-}
-
-bool ChatHandler::HandlePDumpWriteCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    char* file = ExtractQuotedOrLiteralArg(&args);
-    if (!file)
-    {
-        return false;
-    }
-
-    char* p2 = ExtractLiteralArg(&args);
-
-    uint32 lowguid;
-    ObjectGuid guid;
-    // character name can't start from number
-    if (!ExtractUInt32(&p2, lowguid))
-    {
-        std::string name = ExtractPlayerNameFromLink(&p2);
-        if (name.empty())
-        {
-            SendSysMessage(LANG_PLAYER_NOT_FOUND);
-            SetSentErrorMessage(true);
-            return false;
-        }
-
-        guid = sObjectMgr.GetPlayerGuidByName(name);
-        if (!guid)
-        {
-            PSendSysMessage(LANG_PLAYER_NOT_FOUND);
-            SetSentErrorMessage(true);
-            return false;
-        }
-
-        lowguid = guid.GetCounter();
-    }
-    else
-    {
-        guid = ObjectGuid(HIGHGUID_PLAYER, lowguid);
-    }
-
-    if (!sObjectMgr.GetPlayerAccountIdByGUID(guid))
-    {
-        PSendSysMessage(LANG_PLAYER_NOT_FOUND);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    switch (PlayerDumpWriter().WriteDump(file, lowguid))
-    {
-        case DUMP_SUCCESS:
-            PSendSysMessage(LANG_COMMAND_EXPORT_SUCCESS);
-            break;
-        case DUMP_FILE_OPEN_ERROR:
-            PSendSysMessage(LANG_FILE_OPEN_FAIL, file);
-            SetSentErrorMessage(true);
-            return false;
-        default:
-            PSendSysMessage(LANG_COMMAND_EXPORT_FAILED);
-            SetSentErrorMessage(true);
-            return false;
-    }
-
-    return true;
-}
-
 bool ChatHandler::HandleMovegensCommand(char* /*args*/)
 {
     Unit* unit = getSelectedUnit();
@@ -3847,131 +3660,4 @@ bool ChatHandler::HandleModifyGenderCommand(char* args)
     return true;
 }
 
-bool ChatHandler::HandleMmap(char* args)
-{
-    bool on;
-    if (ExtractOnOff(&args, on))
-    {
-        if (on)
-        {
-            sWorld.setConfig(CONFIG_BOOL_MMAP_ENABLED, true);
-            SendSysMessage("WORLD: mmaps are now ENABLED (individual map settings still in effect)");
-        }
-        else
-        {
-            sWorld.setConfig(CONFIG_BOOL_MMAP_ENABLED, false);
-            SendSysMessage("WORLD: mmaps are now DISABLED");
-        }
-        return true;
-    }
-
-    on = sWorld.getConfig(CONFIG_BOOL_MMAP_ENABLED);
-    PSendSysMessage("mmaps are %sabled", on ? "en" : "dis");
-
-    return true;
-}
-
-bool ChatHandler::HandleMmapTestArea(char* args)
-{
-    float radius = 40.0f;
-    ExtractFloat(&args, radius);
-
-    std::list<Creature*> creatureList;
-    MaNGOS::AnyUnitInObjectRangeCheck go_check(m_session->GetPlayer(), radius);
-    MaNGOS::CreatureListSearcher<MaNGOS::AnyUnitInObjectRangeCheck> go_search(creatureList, go_check);
-    // Get Creatures
-    Cell::VisitGridObjects(m_session->GetPlayer(), go_search, radius);
-
-    if (!creatureList.empty())
-    {
-        PSendSysMessage("Found " SIZEFMTD " Creatures.", creatureList.size());
-
-        uint32 paths = 0;
-        uint32 uStartTime = WorldTimer::getMSTime();
-
-        float gx, gy, gz;
-        m_session->GetPlayer()->GetPosition(gx, gy, gz);
-        for (std::list<Creature*>::iterator itr = creatureList.begin(); itr != creatureList.end(); ++itr)
-        {
-            PathFinder path(*itr);
-            path.calculate(gx, gy, gz);
-            ++paths;
-        }
-
-        uint32 uPathLoadTime = WorldTimer::getMSTimeDiff(uStartTime, WorldTimer::getMSTime());
-        PSendSysMessage("Generated %i paths in %i ms", paths, uPathLoadTime);
-    }
-    else
-    {
-        PSendSysMessage("No creatures in %f yard range.", radius);
-    }
-
-    return true;
-}
-
-// use ".mmap testheight 10" selecting any creature/player
-bool ChatHandler::HandleMmapTestHeight(char* args)
-{
-    float radius = 0.0f;
-    ExtractFloat(&args, radius);
-    if (radius > 40.0f)
-    {
-        radius = 40.0f;
-    }
-
-    Unit* unit = getSelectedUnit();
-
-    Player* player = m_session->GetPlayer();
-    if (!unit)
-    {
-        unit = player;
-    }
-
-    if (unit->GetTypeId() == TYPEID_UNIT)
-    {
-        if (radius < 0.1f)
-        {
-            radius = static_cast<Creature*>(unit)->GetRespawnRadius();
-        }
-    }
-    else
-    {
-        if (unit->GetTypeId() != TYPEID_PLAYER)
-        {
-            PSendSysMessage(LANG_SELECT_CHAR_OR_CREATURE);
-            return false;
-        }
-    }
-
-    if (radius < 0.1f)
-    {
-        PSendSysMessage("Provided spawn radius for %s is too small. Using 5.0f instead.", unit->GetGuidStr().c_str());
-        radius = 5.0f;
-    }
-
-    float gx, gy, gz;
-    unit->GetPosition(gx, gy, gz);
-
-    Creature* summoned = unit->SummonCreature(VISUAL_WAYPOINT, gx, gy, gz + 0.5f, 0, TEMPSUMMON_TIMED_DESPAWN, 20000);
-    summoned->CastSpell(summoned, 8599, false);
-    uint32 tries = 1;
-    uint32 successes = 0;
-    uint32 startTime = WorldTimer::getMSTime();
-    for (; tries < 500; ++tries)
-    {
-        unit->GetPosition(gx, gy, gz);
-        if (unit->GetMap()->GetReachableRandomPosition(unit, gx, gy, gz, radius))
-        {
-            unit->SummonCreature(VISUAL_WAYPOINT, gx, gy, gz, 0, TEMPSUMMON_TIMED_DESPAWN, 15000);
-            ++successes;
-            if (successes >= 100)
-            {
-                break;
-            }
-        }
-    }
-    uint32 genTime = WorldTimer::getMSTimeDiff(startTime, WorldTimer::getMSTime());
-    PSendSysMessage("Generated %u valid points for %u try in %ums.", successes, tries, genTime);
-    return true;
-}
